pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sly-slideshow"
version = "0.1.0"
description = "A powerful command-line tool for creating beautiful slideshows from images"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["slideshow", "video", "images", "cli", "moviepy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Graphics",
    "Environment :: Console",
]
dependencies = [
    "moviepy>=1.0.3",
    "pillow>=8.0.0",
    "numpy>=1.20.0",
    "rich>=10.0.0",
    "toml>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "pytest-mock>=3.0.0",
    "hypothesis>=6.0.0",
    "psutil>=5.0.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "mypy>=0.800",
]

[project.urls]
Homepage = "https://github.com/yourusername/sly"
Repository = "https://github.com/yourusername/sly"
Documentation = "https://github.com/yourusername/sly#readme"
"Bug Tracker" = "https://github.com/yourusername/sly/issues"

[project.scripts]
sly = "sly.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sly*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=sly --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 